fix: hangup issue
