+++ apps/rahat-ui/src/sections/projects/aa-2/ivr/ivr.flow.store.ts
@@ -1,10 +1,8 @@
 import { create } from 'zustand';
-import { IvrFlow, IvrFlowNode } from './ivr.flow.types';
+import { IvrFlow, IvrFlowNode, findNodeById } from './ivr.flow.types';
 
 let nodeIdCounter = 0;
 const generateNodeId = () => `node_${++nodeIdCounter}_${Date.now()}`;
-const generateFlowId = () =>
-  `flow_${Date.now()}_${Math.random().toString(36).slice(2, 8)}`;
 
 function createDefaultRootMenu(): IvrFlowNode {
   return {
@@ -17,15 +15,9 @@ function createDefaultRootMenu(): IvrFlowNode {
   };
 }
 
-interface FlowHistory {
-  past: IvrFlow[];
-  future: IvrFlow[];
-}
-
 interface IvrFlowStore {
   flows: IvrFlow[];
   currentFlowId: string | null;
-  history: Record<string, FlowHistory>;
 
   loadFlow: (flowId: string) => IvrFlow | undefined;
   getCurrentFlow: () => IvrFlow | null;
@@ -34,31 +26,6 @@ interface IvrFlowStore {
   updateNode: (nodeId: string, updates: Partial<IvrFlowNode>) => void;
   deleteNode: (nodeId: string) => void;
   setFlowRootMenu: (flowId: string, rootMenu: IvrFlowNode) => void;
-}
-
-function cloneFlow(flow: IvrFlow): IvrFlow {
-  return JSON.parse(JSON.stringify(flow));
-}
-
-function pushHistory(
-  store: IvrFlowStore,
-  flowId: string,
-  currentFlow: IvrFlow,
-) {
-  if (!store.history[flowId]) {
-    store.history[flowId] = { past: [], future: [] };
-  }
-  store.history[flowId].past.push(cloneFlow(currentFlow));
-  store.history[flowId].future = [];
-}
-
-function findNodeById(root: IvrFlowNode, id: string): IvrFlowNode | null {
-  if (root.id === id) return root;
-  for (const child of root.children) {
-    const found = findNodeById(child, id);
-    if (found) return found;
-  }
-  return null;
 }
 
 function removeNodeById(root: IvrFlowNode, id: string): boolean {
@@ -78,7 +45,6 @@ const initialFlows: IvrFlow[] = [];
 export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
   flows: initialFlows,
   currentFlowId: null,
-  history: {},
 
   loadFlow: (flowId) => {
     const { flows } = get();
@@ -97,9 +63,6 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
       return newFlow;
     }
     set({ currentFlowId: flowId });
-    if (!get().history[flowId]) {
-      get().history[flowId] = { past: [], future: [] };
-    }
     return flow;
   },
 
@@ -116,6 +79,9 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
       const parent = findNodeById(flow.rootMenu, parentId);
       if (!parent) return state;
 
+      // Parent now has children, so it should not hangup
+      parent.hangup = false;
+
       const takenDigits = new Set(
         parent.children.map((c) => c.digit).filter((d): d is string => !!d),
       );
@@ -129,8 +95,6 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
 
       if (availableDigit === null) return state;
 
-      pushHistory(state, flow.id, flow);
-
       const newNode: IvrFlowNode = {
         id: generateNodeId(),
         digit: String(availableDigit),
@@ -141,6 +105,9 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
         children: [],
         ...nodePartial,
       };
+
+      // New node has no children, so it should hangup (leaf node)
+      newNode.hangup = true;
 
       parent.children.push(newNode);
 
@@ -160,7 +127,6 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
       const node = findNodeById(flow.rootMenu, nodeId);
       if (!node) return state;
 
-      pushHistory(state, flow.id, flow);
       Object.assign(node, updates, { id: node.id });
 
       return {
@@ -178,8 +144,28 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
 
       if (flow.rootMenu.id === nodeId) return state;
 
-      pushHistory(state, flow.id, flow);
+      // Find parent before deleting the child
+      const findParent = (
+        root: IvrFlowNode,
+        id: string,
+      ): IvrFlowNode | null => {
+        const idx = root.children.findIndex((c) => c.id === id);
+        if (idx !== -1) return root;
+        for (const child of root.children) {
+          const found = findParent(child, id);
+          if (found) return found;
+        }
+        return null;
+      };
+
+      const parent = findParent(flow.rootMenu, nodeId);
+
       removeNodeById(flow.rootMenu, nodeId);
+
+      // If parent has no more children, it becomes a leaf node so should hangup
+      if (parent && parent.children.length === 0) {
+        parent.hangup = true;
+      }
 
       return {
         flows: state.flows.map((f) =>
@@ -193,7 +179,6 @@ export const useIvrFlowStore = create<IvrFlowStore>((set, get) => ({
     set((state) => {
       const flow = state.flows.find((f) => f.id === flowId);
       if (!flow) return state;
-      pushHistory(state, flow.id, flow);
       flow.rootMenu = JSON.parse(JSON.stringify(rootMenu));
       return {
         flows: state.flows.map((f) =>
